deps/v8/test/inspector/debugger/provisional-breakpoint-for-anonymous-script-expected.txt

Checks provisional breakpoints by hash in anonymous scripts

Running test: testNextScriptParsed
function foo(){#}


Running test: testPreviousScriptParsed
var list = list ? list.concat(foo) : [foo]; function foo(){#}

var list = list ? list.concat(foo) : [foo]; function foo(){#}
